django_app/requirements.txt
django>=5,<6
django-compressor>=4.5
django-htmlmin~=0.11
djangorestframework>=3.10
html5validator~=0.4
psycopg2-binary~=2.9
# mistune>=2.0.3 # not directly required, pinned by Snyk to avoid a vulnerability
# drf-spectacular[sidecar] # for schema generation
# uritemplate  # for python manage.py generateschema
# django-filter
# django-crispy-forms  # needed by django_filters/rest_framework
# crispy-bootstrap5
